LSystemGenerator Done; Change in LSystemManager;
Processing of the sentence in LSystemGenerator; Function MergeLSystem in LSystemManager should be done. Not tested yet.

+++ L-System_Roads/Assets/Scripts/Visualizer/AgentParameter.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class AgentParameter
 {
     public Vector3 position, direction;
-    public int length;
+    public float length;
 }
